src/requirements.txt
absl-py==1.4.0
accelerate
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
alembic==1.10.3
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
asttokens==2.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blis==0.7.9
boto3==1.26.106
botocore==1.29.106
build==0.10.0
cachetools==5.2.0
catalogue==2.0.8
catboost==1.2
category-encoders==2.6.0
certifi==2022.09.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.0
cmaes==0.9.1
colorama==0.4.6
colorful==0.5.4
colorlog==6.7.0
confection==0.0.4
contextvars==2.4
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
datasets==2.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
dm-tree==0.1.8
docutils==0.20.1
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
etils==1.3.0
evaluate==0.3.0
exceptiongroup==1.1.0
executing==1.2.0
fairlearn==0.8.0
fairscale
fastai==2.7.12
fastapi==0.88.0
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.16.2
fastprogress==1.0.3
fasttext==0.9.1
filelock==3.8.0
folktables
fonttools==4.38.0
frozendict==2.3.4
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gdown==4.7.1
git+https://github.com/sxjscience/automl_multimodal_benchmark.git#subdirectory=multimodal_text_benchmark
gluonts==0.12.6
google-api-core==2.10.2
google-auth==2.17.2
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.57.0
gpustat==1.0.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.13.3
hyperopt==0.2.7
idna==3.4
imageio==2.27.0
immutables==0.19
importlib-metadata==5.1.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
jaraco.classes==3.2.3
javalang==0.13.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonlines==3.1.0
jsonschema==4.17.3
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kaggle==1.5.13
keyring==23.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lightgbm-ray==0.1.8
lightning-utilities==0.8.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
mdurl==0.1.2
mistune==2.0.4
model-index==0.1.11
more-itertools==9.1.0
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nlpaug==1.1.11
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
nptyping==2.4.1
numba
numpy
nvidia-ml-py==11.495.46
oauthlib==3.2.2
omegaconf==2.2.3
opencensus==0.11.0
opencensus-context==0.1.3
openmim==0.3.7
openpyxl==3.1.2
optuna==3.1.1
ordered-set==4.1.0
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
plotly==5.14.0
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.2
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==1.9.4
pytorch-metric-learning==1.7.3
pytorch-tabnet==4.0
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
ray==2.2.0
readme-renderer==37.3
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3986==2.0.0
rich==13.3.3
rsa==4.9
rtdl==0.0.13
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
sentencepiece==0.1.97
setuptools==64
seqeval==1.2.2
six==1.16.0
scikit-learn==1.1.3
smart-open==6.2.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.7.2
SQLAlchemy==2.0.9
srsly==2.4.6
stack-data==0.6.2
starlette==0.22.0
statsforecast==1.4.0
statsmodels
tab-transformer-pytorch==0.2.1
tabulate==0.9.0
tbats==1.1.2
tenacity==8.2.2
tensorflow-datasets==4.9.2
tensorflow-metadata==1.13.1
termcolor==2.3.0
terminado==0.17.0
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torchinfo==1.8.0
torchmetrics==0.8.2
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
transformers==4.26.1
tune-sklearn==0.4.5
twine==4.0.2
typer==0.7.0
typing_extensions==4.4.0
ujson==5.7.0
unidiff==0.7.5
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.0
wasabi==1.1.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
wrapt==1.14.1
xarray==2023.1.0
xgboost==1.7.1
xgboost-ray==0.1.12
xport
xxhash==3.2.0
yacs==0.1.8
yarl==1.8.1
zipp==3.11.0
